requirements.txt
absolute==1.0.0
adyen==5.1.0
aldryn-forms==6.2.1
aldryn-search==1.1.0
aniso8601==7.0.0
ariadne==0.11.0
black==21.7b0
bleach==4.0.0
brotli==1.0.9
celery==5.1.2
channels==3.0.4
dj-email-url==1.0.2
django-admin-sortable2==1.0
django-app-enabler==0.2.0
django-bootstrap3==15.0.0
django-cache-url==3.2.3
django-cors-headers==3.7.0
django-enumfields==2.1.1
django-environ==0.4.5
django-extensions==3.1.3
django-gdpr-assist==1.3.0
django-import-export==2.5.0
django-jazzmin==2.4.8
django-jinja==2.9.0
django-measurement==3.2.3
django-newsletter==0.9.1
django-oscar-api==3.0.0
django-oscar-invoices==0.2
django-oscar-paypal==2.0.0
django-otp-yubikey==1.0.0.post1
django-payments==0.14.0
django-photologue==3.13
django-redis==5.0.0
django-registration-redux==2.9
django-reversion==4.0.0
django-shortuuidfield==0.1.3
django-simple-captcha==0.5.14
django-slick-reporting==0.6.0
django-timezone-field==4.2.1
django-two-factor-auth==1.13.1
djangocms-audio==2.0.0
djangocms-blog==1.2.3
djangocms-bootstrap4==2.0.0
djangocms-charts==3.0.2
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-installer==2.0.0
djangocms-maps==0.13.0
djangocms-named-menus==2.0.1
djangocms-page-meta==1.0.1
djangocms-style==3.0.0
djangocms-video==3.0.0
fake-factory==9999.9.9
flake8-debugger==4.0.0
google-api-python-client==1.4.1
graphene-django==2.0.0
graphqlclient==0.2.4
gunicorn==20.1.0
ipdb==0.13.9
isort==5.9.3
jaeger-client==4.6.0
jdcal==1.4.1
keyring==23.1.0
keyrings.alt==4.1.0
ldap3==2.9.1
markuppy==1.14
mixer==7.1.2
mock==4.0.3
mysql-connector-python==8.0.26
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
paypal==1.2.5
pinax-announcements==4.0.0
pinax-badges==3.0.0
pinax-calendars==3.0.0
pinax-events==3.0.0
pinax-messages==3.0.0
pinax-testimonials==3.0.0
pip-chill==1.0.1
puput==1.1.1
pycountry==20.7.3
pyinstaller==4.5.1
pyopenssl==20.0.1
pypiwin32==223
pyprof2calltree==1.4.5
pysolr==3.9.0
pytest-cache==1.0
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-splinter==3.3.1
python-keycloak-client==0.2.3
pytimeparse==1.1.8
pyxb==1.2.6
rx==1.6.1
sentry-sdk==1.3.1
service-identity==21.1.0
sphinx==4.1.2
supervisor==4.2.2
supervisor-win
twilio==6.63.0
virtualenv==20.7.2
weasyprint==53.0
werkzeug==2.0.1
wfastcgi==3.0.0
whitenoise==5.3.0
whoosh==2.7.4
wq==1.3.0b1
xlwt==1.3.0
zopfli==0.1.8